backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==9.0.0
bcrypt==4.1.2
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
insightface-0.7.3
onnxruntime>=1.16.0
